API

This page contains documentation of the public API of the RegularizedLeastSquares. In the Julia REPL one can access this documentation by entering the help mode with ?

General

RegularizedLeastSquares.linearSolverList — Function

Return a list of all available linear solvers

RegularizedLeastSquares.createLinearSolver — Function

createLinearSolver(solver::AbstractLinearSolver, A; log::Bool=false, kargs...)

This method creates a solver. The supported solvers are methods typically used for solving regularized linear systems. All solvers return an approximate solution to Ax = b.

TODO: give a hint what solvers are available

RegularizedLeastSquares.solve — Method

solve(solver::AbstractLinearSolver, b; solverInfo = nothing, kwargs...)

solves an inverse problem using solver. When solverInfo is passed to the function information about the iteration process is stored.

See also SolverInfo.

RegularizedLeastSquares.SolverInfo — Type

The Solver Info type is used to obtain aditional information of linear solvers and their iteration process.

Fields

convMeas::Vector{Float64} - Vector with convergence parameters

nrmse::Vector{Float64} - NRMSD of the current iterate w.r.t some reference (if provided)

x_ref::Vector{T} - Reference for computing NRMSD of iterates

x_iter::Vector{Vector{T}} - Vector for storing iterates

store_solutions::Bool - specification whether iterates are to be stored

the content of convMeas is specific to each solver. It can be found in the documentation of the corresping solve-method.

RegularizedLeastSquares.applicableSolverList — Function

applicable(args...)

list all solvers that are applicable to the given arguments. Arguments are the same as for isapplicable without the solver type.

See also isapplicable, linearSolverList.

RegularizedLeastSquares.isapplicable — Function

isapplicable(solverType::Type{<:AbstractLinearSolver}, A, x, reg)

return true if a solver of type solverType is applicable to system matrix A, data x and regularization terms reg.

ADMM

RegularizedLeastSquares.ADMM — Type

ADMM(A, x; kwargs...)

creates an ADMM object for the system matrix A.

Arguments

A - system matrix

x - (optional) array with the same type and size as the solution

Keywords

reg - regularization term vector

normalizeReg - regularization normalization scheme

precon=Identity() - preconditionner for the internal CG algorithm

ρ::Real=1.e-2 - penalty of the augmented lagrangian

vary_ρ::Bool=:none - vary rho to balance primal and dual feasibility

iterations::Int64=50 - max number of ADMM iterations

iterationsInner::Int64=10 - max number of internal CG iterations

absTol::Real=eps() - abs tolerance for stopping criterion

relTol::Real=eps() - rel tolerance for stopping criterion

tolInner::Real=1.e-5 - rel tolerance for CG stopping criterion

See also createLinearSolver, solve.

RegularizedLeastSquares.solve — Method

solve(solver::ADMM, b; kwargs...) where {matT,vecT}

solves an inverse problem using ADMM.

Arguments

solver::ADMM - the solver containing both system matrix and regularizer

b::Vector - data vector

Keywords

A::matT=solver.A - operator for the data-term of the problem

startVector::Vector{T}=T[] - initial guess for the solution

solverInfo=nothing - solverInfo for logging

when a SolverInfo objects is passed, the primal residuals solver.rᵏ and the dual residual norm(solver.sᵏ) are stored in solverInfo.convMeas.

CGNR

RegularizedLeastSquares.CGNR — Type

CGNR(A, x; kargs...)

creates an CGNR object for the system matrix A.

Arguments

A - system matrix

x::vecT - (optional) array with the same type and size as the solution

Keywords

reg - regularization term vector

normalizeReg - regularization normalization scheme

weights::vecT=eltype(A)[] - weights for the data term

AᴴA=A'*A - specialized normal operator, default is A'*A

iterations::Int64=10 - number of iterations

relTol::Float64=eps() - rel tolerance for stopping criterion

See also createLinearSolver, solve.

RegularizedLeastSquares.solve — Method

solve(solver::CGNR, u; kwargs...) where vecT

solves Tikhonov-regularized inverse problem using CGNR.

Arguments

`solver::CGNR - the solver containing both system matrix and regularizer

u::vecT - data vector

Keywords

startVector::vecT=similar(u,0) - initial guess for the solution

solverInfo=nothing - solverInfo for logging

when a SolverInfo objects is passed, the residuals solver.zl are stored in solverInfo.convMeas.

Kaczmarz

RegularizedLeastSquares.Kaczmarz — Type

Kaczmarz(S, b; kwargs...)

creates a Kaczmarz object

Arguments

S - system matrix

b=nothing - measurement

Keywords

reg - regularization term vector

normalizeReg - regularization normalization scheme

weights::Vector{R}=ones(Float64,size(S,1)) - weights for the data term

enforceReal::Bool=false - constrain the solution to be real

enforcePositive::Bool=false - constrain the solution to have positive real part

randomized::Bool=false - randomize Kacmarz algorithm

subMatrixFraction::Float64=0.1 - fraction of rows used in randomized Kaczmarz algorithm

shuffleRows::Bool=false - randomize Kacmarz algorithm

seed::Int=1234 - seed for randomized algorithm

iterations::Int64=10 - number of iterations

See also createLinearSolver, solve.

RegularizedLeastSquares.solve — Method

solve(solver::AbstractLinearSolver, b; solverInfo = nothing, kwargs...)

solves an inverse problem using solver. When solverInfo is passed to the function information about the iteration process is stored.

See also SolverInfo.

solve(solver::Kaczmarz, u::Vector; kwargs...)

solves Tikhonov-regularized inverse problem using Kaczmarz algorithm.

Arguments

`solver::Kaczmarz - the solver containing both system matrix and regularizer

u::Vector - data vector

Keywords

S::matT=solver.S - operator for the data-term of the problem

startVector::Vector{T}=T[] - initial guess for the solution

weights::Vector{T}=solver.weights - weights for the data term

shuffleRows::Bool=false - randomize Kacmarz algorithm

solverInfo=nothing - solverInfo for logging

when a SolverInfo objects is passed, the residuals are stored in solverInfo.convMeas.

FISTA

RegularizedLeastSquares.FISTA — Type

FISTA(A, x; kwargs...)

creates a FISTA object for the system matrix A.

Arguments

A - system matrix

x::vecT - (optional) array with the same type and size as the solution

Keywords

reg - regularization term vector

normalizeReg - regularization normalization scheme

AᴴA=A'*A - specialized normal operator, default is A'*A

ρ=0.95 - step size for gradient step

normalize_ρ=true - normalize step size by the maximum eigenvalue of AᴴA

t=1.0 - parameter for predictor-corrector step

relTol::=eps(real(T)) - tolerance for stopping criterion

iterations::Int64=50 - maximum number of iterations

restart::Symbol=:none - :none, :gradient options for restarting

See also createLinearSolver, solve.

RegularizedLeastSquares.solve — Method

solve(solver::FISTA, b; kwargs...)

solves an inverse problem using FISTA.

Arguments

solver::FISTA - the solver containing both system matrix and regularizer

b::vecT - data vector

Keywords

A=solver.A - operator for the data-term of the problem

startVector::vecT=similar(b,0) - initial guess for the solution

solverInfo=nothing - solverInfo object

when a SolverInfo objects is passed, the residuals are stored in solverInfo.convMeas.

OptISTA

RegularizedLeastSquares.OptISTA — Type

OptISTA(A, x; kwargs...)

creates a OptISTA object for the system matrix A. OptISTA has a 2x better worst-case bound than FISTA, but actual performance varies by application. It stores 2 extra intermediate variables the size of the image compared to FISTA

Reference:

Uijeong Jang, Shuvomoy Das Gupta, Ernest K. Ryu, "Computer-Assisted Design of Accelerated Composite Optimization Methods: OptISTA," arXiv:2305.15704, 2023, [https://arxiv.org/abs/2305.15704]

Arguments

A - system matrix

x::vecT - array with the same type and size as the solution

Keywords

reg - regularization term vector

normalizeReg - regularization normalization scheme

AᴴA=A'*A - specialized normal operator, default is A'*A

ρ=0.95 - step size for gradient step

normalize_ρ=true - normalize step size by the maximum eigenvalue of AᴴA

θ=1.0 - parameter for predictor-corrector step

relTol::Float64=1.e-5 - tolerance for stopping criterion

iterations::Int64=50 - maximum number of iterations

See also createLinearSolver, solve.

RegularizedLeastSquares.solve — Method

solve(solver::OptISTA, b::Vector; kwargs...)

solves an inverse problem using OptISTA.

Arguments

solver::OptISTA - the solver containing both system matrix and regularizer

b::vecT - data vector

Keywords

A=solver.A - operator for the data-term of the problem

startVector::vecT=similar(b,0) - initial guess for the solution

solverInfo=nothing - solverInfo object

when a SolverInfo objects is passed, the residuals are stored in solverInfo.convMeas.

POGM

RegularizedLeastSquares.POGM — Type

POGM(A, x; kwargs...)

creates a POGM object for the system matrix A. POGM has a 2x better worst-case bound than FISTA, but actual performance varies by application. It stores 3 extra intermediate variables the size of the image compared to FISTA Only gradient restart scheme is implemented for now

References:

A.B. Taylor, J.M. Hendrickx, F. Glineur, "Exact worst-case performance of first-order algorithms for composite convex optimization," Arxiv:1512.07516, 2015, SIAM J. Opt. 2017 [http://doi.org/10.1137/16m108104x]

Kim, D., & Fessler, J. A. (2018). Adaptive Restart of the Optimized Gradient Method for Convex Optimization. Journal of Optimization Theory and Applications, 178(1), 240–263. [https://doi.org/10.1007/s10957-018-1287-4]

Arguments

A - system matrix

x::vecT - array with the same type and size as the solution

Keywords

reg - regularization term vector

normalizeReg - regularization normalization scheme

AᴴA=A'*A - specialized normal operator, default is A'*A

ρ=0.95 - step size for gradient step

normalize_ρ=true - normalize step size by the maximum eigenvalue of AᴴA

t=1 - parameter for predictor-corrector step

σ_fac=1 - parameter for decreasing γ-momentum ∈ [0,1]

relTol::=eps(real(T)) - tolerance for stopping criterion

iterations::Int64=50 - maximum number of iterations

restart::Symbol=:none - :none, :gradient options for restarting

verbose::Bool=false - print residual norm in each iteration

See also createLinearSolver, solve.

RegularizedLeastSquares.solve — Method

solve(solver::POGM, b::Vector)

solves an inverse problem using POGM.

Arguments

solver::POGM - the solver containing both system matrix and regularizer

b::vecT - data vector

Keywords

A=solver.A - operator for the data-term of the problem

startVector::vecT=similar(b,0) - initial guess for the solution

solverInfo=nothing - solverInfo object

when a SolverInfo objects is passed, the residuals are stored in solverInfo.convMeas.

SplitBregman

RegularizedLeastSquares.SplitBregman — Type

SplitBregman(A, x; kwargs...)

creates a SplitBregman object for the system matrix A.

Arguments

A::matT - system matrix

x::vecT - Array with the same type and size as the solution

Keywords

reg=nothing - regularization term vector

normalizeReg - regularization normalization scheme

precon=Identity() - preconditionner for the internal CG algorithm

ρ=[1.e2] - weights for condition on regularized variables

iterations::Int64=10 - number of outer iterations

iterationsInner::Int64=50 - maximum number of inner iterations

iterationsCG::Int64=10 - maximum number of CG iterations

absTol::Float64=eps() - abs tolerance for stopping criterion

relTol::Float64=eps() - rel tolerance for stopping criterion

tolInner::Float64=1.e-5 - tolerance for CG stopping criterion

See also createLinearSolver, solve.

RegularizedLeastSquares.solve — Method

solve(solver::AbstractLinearSolver, b; solverInfo = nothing, kwargs...)

solves an inverse problem using solver. When solverInfo is passed to the function information about the iteration process is stored.

See also SolverInfo.

solve(solver::SplitBregman, b; kwargs...)

solves an inverse problem using the Split Bregman method.

Arguments

solver::SplitBregman - the solver containing both system matrix and regularizer

b::vecT - data vector

Keywords

A::matT=solver.A - operator for the data-term of the problem

startVector::vecT=similar(b,0) - initial guess for the solution

solverInfo=nothing - solverInfo for logging

when a SolverInfo objects is passed, the primal residuals solver.rk and the dual residual norm(solver.sk) are stored in solverInfo.convMeas.

Direct