API

Solvers

RegularizedLeastSquares.linearSolverList — Function

Return a list of all available linear solvers

RegularizedLeastSquares.createLinearSolver — Function

createLinearSolver(solver::AbstractString, A; log::Bool=false, kargs...)

This method creates a solver. The supported solvers are methods typically used in MPI/MRI. All solvers return an approximate solution to STx = u. Function returns choosen solver.

solvers:

"kaczmarz" - kaczmarz method (the default)

"cgnr - CGNR

"direct" - A direct solver using the backslash operator

"daxkaczmarz" - Dax algorithm (with Kaczmarz) for unconstrained problems

"daxconstrained" - Dax algorithm for constrained problems

"pseudoinverse" - approximates a solution using the More-Penrose pseudo inverse

"fusedlasso" - solver for the Fused-Lasso problem

"fista" - Fast Iterative Shrinkage Thresholding Algorithm

"admm" - Alternating Direcion of Multipliers Method

"splitBregman" - Split Bregman method for constrained & regularized inverse problems

"primaldualsolver"- First order primal dual method

Regularization

RegularizedLeastSquares.Regularization — Type

Type describing Regularizers

Fields

prox!::Function - proximal map for the regularizer

norm::Function - (semi-)norm for the regularizer

λ::AbstractFloat - regularization paramter

params::Dict{Symbol,Any} - additional parameters

RegularizedLeastSquares.Regularization — Method

Regularization(name::String, λ::AbstractFloat; kargs...)

create a Regularization object containing all the infos necessary to calculate a proximal map.

valid names

"L2"

"L1"

"L21"

"TV"

"LLR"

"Nuclear"

"Positive"

"Proj"

RegularizedLeastSquares.Regularization — Method

Regularization(names::Vector{String}, λ::Vector{Float64}; kargs...)

create a Regularization object containing all the infos necessary to calculate a proximal map. Valid names are the same as in Regularization(name::String, λ::AbstractFloat; kargs...).

RegularizedLeastSquares.RegularizationList — Function

RegularizationList()

Returns a list of all available Regularizations

Missing docstring.

Missing docstring for RegularizedLeastSquares.normalize!. Check Documenter's build log for details.

RegularizedLeastSquares.proxL1! — Function

proxL1!(x::Array{T}, λ::Float64; sparseTrafo::Trafo=nothing, kargs...) where T

performs soft-thresholding - i.e. proximal map for the Lasso problem.

Arguments:

x::Array{T} - Vector to apply proximal map to

λ::Float64 - regularization paramter

sparseTrafo::Trafo=nothing - sparsifying transform to apply

RegularizedLeastSquares.proxL2! — Function

proxL2!(x::Vector{T}, λ::Float64; kargs...) where T

proximal map for Tikhonov regularization.

RegularizedLeastSquares.proxL21! — Function

proxL21!(x::Vector{T},λ::Float64; sparseTrafo::Trafo=nothing, slices::Int64=1, kargs...)

group-soft-thresholding for l1/l2-regularization.

Arguments:

x::Array{T} - Vector to apply proximal map to

λ::Float64 - regularization paramter

sparseTrafo::Trafo=nothing - sparsifying transform to apply

slices::Int64=1 - number of elements per group

RegularizedLeastSquares.proxLLR! — Function

proxLLR!(x::Vector{T}, λ::Float64=1e-6; kargs...) where T

proximal map for LLR regularization using singular-value-thresholding

Arguments

x::Vector{T} - Vector to apply proximal map to

λ::Float64 - regularization parameter

shape::Tuple{Int}=[] - dimensions of the image

blockSize::Tuple{Int}=[2;2] - size of patches to perform singluar value thresholding on

randshift::Bool=true - randomly shifts the patches to ensure translation invariance

RegularizedLeastSquares.proxNuclear! — Function

proxNuclear!(x::Vector{T}, λ::Float64; svtShape::NTuple=[],kargs...)

applies singular value soft-thresholding - i.e. the proximal map for the nuclear norm regularization.

Arguments:

x::Array{T} - Vector to apply proximal map to

λ::Float64 - regularization paramter

svtShape::NTuple=[] - size of the underlying matrix

RegularizedLeastSquares.proxPositive! — Function

proxPositive!(x::Vector{T},λ::Float64=1.0;kargs...) where T

enforce positivity and realness of solution `x`.

RegularizedLeastSquares.proxProj! — Function

proxProj!(x::Vector{T}, λ::Float64; projFunc=x->x, kargs...)

applies the projection given by projFunc.

RegularizedLeastSquares.proxTV! — Function

proxTV!(x::Vector{T}, λ::Float64; kargs...) where T

proximal map for ansitropic TV regularization using the Fast Gradient Projection algorithm.

Arguments

x::Array{T} - Vector to apply proximal map to

λ::Float64 - regularization paramter

shape::NTuple=[] - size of the underlying image

iterationsTV::Int64=20 - number of FGP iterations

weights::Array=[] - weights to apply to the image gradients

RegularizedLeastSquares.normL1 — Function

normL1(x::Array{T}, λ::Float64; sparseTrafo::Trafo=nothing, kargs...) where T

returns the value of the L1-regularization term. Arguments are the same as in proxL1!

RegularizedLeastSquares.normL2 — Function

normL2(x::Vector{T}, λ::Float64, kargs...)

returns the value of the L2-regularization term

RegularizedLeastSquares.normL21 — Function

normL21(x::Vector{T}, λ::Float64; sparseTrafo::Trafo=nothing, slices::Int64=1, kargs...) where T

return the value of the L21-regularization term. Arguments are the same as in proxL21!

RegularizedLeastSquares.normLLR — Function

normLLR(x::Vector{T}, λ::Float64; kargs...) where T

returns the value of the LLR-regularization term. Arguments are the same is in proxLLR!

RegularizedLeastSquares.normNuclear — Function

normNuclear(x::Vector{T}, λ::Float64; svtShape::NTuple=[],kargs...) where T

returns the value of the nuclear norm regularization term. Arguments are the same as in proxNuclear!

RegularizedLeastSquares.normPositive — Function

returns the value of the characteristic function of real, positive numbers.
normPositive(x) = (isreal(x)&&x>0) ? 0 : Inf

RegularizedLeastSquares.normProj — Function

normProj(x::Vector{T}, λ::Float64=0.0; projFunc=x->x, kargs...) where T

evaluate indicator function of set to be projected onto.

RegularizedLeastSquares.normTV — Function

normTV(x::Vector{T},λ::Float64;shape::NTuple=[],kargs...) where T

returns the value of the ansisotropic TV-regularization term. Arguments are the same as in proxTV!

LinearOperators

Missing docstring.

Missing docstring for RegularizedLeastSquares.linearOperator(op::AbstractString, shape). Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegularizedLeastSquares.linearOperatorList. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegularizedLeastSquares.DCTOp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegularizedLeastSquares.DSTOp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegularizedLeastSquares.FFTOp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegularizedLeastSquares.SamplingOp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegularizedLeastSquares.WaveletOp. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegularizedLeastSquares.WeightingOp. Check Documenter's build log for details.