API for Solvers

This page contains documentation of the public API of the RegularizedLeastSquares. In the Julia REPL one can access this documentation by entering the help mode with ?

solve

RegularizedLeastSquares.solve — Method

solve(solver::AbstractLinearSolver, b; x0 = 0, f_trace = (_, _) -> nothing)

Solves an inverse problem for the data vector b using solver.

Required Arguments

solver::AbstractLinearSolver - linear solver (e.g., ADMM or FISTA), containing forward/normal operator and regularizer

b::AbstractVector - data vector if A was supplied to the solver, back-projection of the data otherwise

Optional Keyword Arguments

x0::AbstractVector - initial guess for the solution; default is zero

f_trace::Function - function that takes the two arguments f_trace(solver, iteration) and, e.g., stores, prints, or plots the intermediate solutions or convergence parameters.

Examples

The optimization problem

\[ argmin_x ||Ax - b||_2^2 + λ ||x||_1\]

can be solved with the following lines of code:

julia> using RegularizedLeastSquares

julia> A = [0.831658  0.96717
            0.383056  0.39043
            0.820692  0.08118];

julia> x = [0.5932234523399985; 0.2697534345340015];

julia> b = A * x;

julia> S = ADMM(A);

julia> x_approx = solve(S, b)
2-element Vector{Float64}:
 0.5932234523399984
 0.26975343453400163

Here, we use L1Regularization, which is default for ADMM. All regularization options can be found in API for Regularizers.

The following example solves the same problem, but stores the solution x of each interation in tr:

julia> tr = Dict[]
Dict[]

julia> store_trace!(tr, solver, iteration) = push!(tr, Dict("iteration" => iteration, "x" => solver.x, "beta" => solver.β))
store_trace! (generic function with 1 method)

julia> x_approx = solve(S, b; f_trace=(solver, iteration) -> store_trace!(tr, solver, iteration))
2-element Vector{Float64}:
 0.5932234523399984
 0.26975343453400163

julia> tr[3]
Dict{String, Any} with 3 entries:
  "iteration" => 2
  "x"         => [0.593223, 0.269753]
  "beta"      => [1.23152, 0.927611]

The last example show demonstrates how to plot the solution at every 10th iteration:

julia> using Plots

julia> function plot_trace(solver, iteration)
         if iteration % 10 == 0
           display(scatter(solver.x))
         end
       end
plot_trace (generic function with 1 method)

julia> x_approx = solve(S, b; f_trace = plot_trace);

The keyword f_trace allows you to pass any function that takes the arguments solver and iteration and prints, stores, or plots intermediate result.

ADMM

RegularizedLeastSquares.ADMM — Type

ADMM(A; AHA = A'*A, precon = Identity(), reg = L1Regularization(zero(eltype(AHA))), normalizeReg = NoNormalization(), rho = 1e-1, vary_rho = :none, iterations = 50, iterationsInner = 10, absTol = eps(real(eltype(AHA))), relTol = eps(real(eltype(AHA))), tolInner = 1e-5, verbose = false)
ADMM( ; AHA = ,     precon = Identity(), reg = L1Regularization(zero(eltype(AHA))), normalizeReg = NoNormalization(), rho = 1e-1, vary_rho = :none, iterations = 50, iterationsInner = 10, absTol = eps(real(eltype(AHA))), relTol = eps(real(eltype(AHA))), tolInner = 1e-5, verbose = false)

creates an ADMM object for the forward operator A or normal operator AHA.

Required Arguments

A - forward operator

OR

AHA - normal operator (as a keyword argument)

Optional Keyword Arguments

AHA - normal operator is optional if A is supplied

precon - preconditionner for the internal CG algorithm

reg::AbstractParameterizedRegularization - regularization term; can also be a vector of regularization terms

normalizeReg::AbstractRegularizationNormalization - regularization normalization scheme; options are NoNormalization(), MeasurementBasedNormalization(), SystemMatrixBasedNormalization()

rho::Real - penalty of the augmented Lagrangian

vary_rho::Symbol - vary rho to balance primal and dual feasibility; options :none, :balance, :PnP

iterations::Int - maximum number of (outer) ADMM iterations

iterationsInner::Int - max number of (inner) CG iterations

absTol::Real - abs tolerance for stopping criterion

relTol::Real - tolerance for stopping criterion

tolInner::Real - rel tolerance for CG stopping criterion

verbose::Bool - print residual in each iteration

See also createLinearSolver, solve.

CGNR

RegularizedLeastSquares.CGNR — Type

CGNR(A; AHA = A' * A, reg = L2Regularization(zero(real(eltype(AHA)))), normalizeReg = NoNormalization(), weights = similar(AHA, 0), iterations = 10, relTol = eps(real(eltype(AHA))))
CGNR( ; AHA = ,       reg = L2Regularization(zero(real(eltype(AHA)))), normalizeReg = NoNormalization(), weights = similar(AHA, 0), iterations = 10, relTol = eps(real(eltype(AHA))))

creates an CGNR object for the forward operator A or normal operator AHA.

Required Arguments

A - forward operator

OR

AHA - normal operator (as a keyword argument)

Optional Keyword Arguments

AHA - normal operator is optional if A is supplied

reg::AbstractParameterizedRegularization - regularization term; can also be a vector of regularization terms

normalizeReg::AbstractRegularizationNormalization - regularization normalization scheme; options are NoNormalization(), MeasurementBasedNormalization(), SystemMatrixBasedNormalization()

weights::AbstactVector - weights for the data term; must be of same length and type as the data term

iterations::Int - maximum number of iterations

relTol::Real - tolerance for stopping criterion

See also createLinearSolver, solve.

Kaczmarz

RegularizedLeastSquares.Kaczmarz — Type

Kaczmarz(A; reg = L2Regularization(0), normalizeReg = NoNormalization(), weights=nothing, randomized=false, subMatrixFraction=0.15, shuffleRows=false, seed=1234, iterations=10, regMatrix=nothing)

Creates a Kaczmarz object for the forward operator A.

Required Arguments

A - forward operator

Optional Keyword Arguments

reg::AbstractParameterizedRegularization - regularization term

normalizeReg::AbstractRegularizationNormalization - regularization normalization scheme; options are NoNormalization(), MeasurementBasedNormalization(), SystemMatrixBasedNormalization()

weights::AbstractVector - weights for the data term

randomized::Bool - randomize Kacmarz algorithm

subMatrixFraction::Real - fraction of rows used in randomized Kaczmarz algorithm

shuffleRows::Bool - randomize Kacmarz algorithm

seed::Int - seed for randomized algorithm

iterations::Int - number of iterations

See also createLinearSolver, solve.

FISTA

RegularizedLeastSquares.FISTA — Type

FISTA(A; AHA=A'*A, reg=L1Regularization(zero(eltype(AHA))), normalizeReg=NoNormalization(), rho=0.95, normalize_rho=true, theta=1, relTol=eps(real(eltype(AHA))), iterations=50, restart = :none, verbose = false)
FISTA( ; AHA=,     reg=L1Regularization(zero(eltype(AHA))), normalizeReg=NoNormalization(), rho=0.95, normalize_rho=true, theta=1, relTol=eps(real(eltype(AHA))), iterations=50, restart = :none, verbose = false)

creates a FISTA object for the forward operator A or normal operator AHA.

Required Arguments

A - forward operator

OR

AHA - normal operator (as a keyword argument)

Optional Keyword Arguments

AHA - normal operator is optional if A is supplied

precon - preconditionner for the internal CG algorithm

reg::AbstractParameterizedRegularization - regularization term; can also be a vector of regularization terms

normalizeReg::AbstractRegularizationNormalization - regularization normalization scheme; options are NoNormalization(), MeasurementBasedNormalization(), SystemMatrixBasedNormalization()

rho::Real - step size for gradient step

normalize_rho::Bool - normalize step size by the largest eigenvalue of AHA

theta::Real - parameter for predictor-corrector step

relTol::Real - tolerance for stopping criterion

iterations::Int - maximum number of iterations

restart::Symbol - :none, :gradient options for restarting

verbose::Bool - print residual in each iteration

See also createLinearSolver, solve.

OptISTA

RegularizedLeastSquares.OptISTA — Type

OptISTA(A; AHA=A'*A, reg=L1Regularization(zero(eltype(AHA))), normalizeReg=NoNormalization(), rho=0.95, normalize_rho=true, theta=1, relTol=eps(real(eltype(AHA))), iterations=50, verbose = false)
OptISTA( ; AHA=,     reg=L1Regularization(zero(eltype(AHA))), normalizeReg=NoNormalization(), rho=0.95, normalize_rho=true, theta=1, relTol=eps(real(eltype(AHA))), iterations=50, verbose = false)

creates a OptISTA object for the forward operator A or normal operator AHA. OptISTA has a 2x better worst-case bound than FISTA, but actual performance varies by application. It stores 2 extra intermediate variables the size of the image compared to FISTA.

Reference:

Uijeong Jang, Shuvomoy Das Gupta, Ernest K. Ryu, "Computer-Assisted Design of Accelerated Composite Optimization Methods: OptISTA," arXiv:2305.15704, 2023, [https://arxiv.org/abs/2305.15704]

Required Arguments

A - forward operator

OR

AHA - normal operator (as a keyword argument)

Optional Keyword Arguments

AHA - normal operator is optional if A is supplied

reg::AbstractParameterizedRegularization - regularization term

normalizeReg::AbstractRegularizationNormalization - regularization normalization scheme; options are NoNormalization(), MeasurementBasedNormalization(), SystemMatrixBasedNormalization()

rho::Real - step size for gradient step

normalize_rho::Bool - normalize step size by the largest eigenvalue of AHA

theta::Real - parameter for predictor-corrector step

relTol::Real - tolerance for stopping criterion

iterations::Int - maximum number of iterations

verbose::Bool - print residual in each iteration

See also createLinearSolver, solve.

POGM

RegularizedLeastSquares.POGM — Type

POGM(A; AHA = A'*A, reg = L1Regularization(zero(eltype(AHA))), normalizeReg = NoNormalization(), rho = 0.95, normalize_rho = true, theta = 1, sigma_fac = 1, relTol = eps(real(eltype(AHA))), iterations = 50, restart = :none, verbose = false)
POGM( ; AHA = ,     reg = L1Regularization(zero(eltype(AHA))), normalizeReg = NoNormalization(), rho = 0.95, normalize_rho = true, theta = 1, sigma_fac = 1, relTol = eps(real(eltype(AHA))), iterations = 50, restart = :none, verbose = false)

Creates a POGM object for the forward operator A or normal operator AHA. POGM has a 2x better worst-case bound than FISTA, but actual performance varies by application. It stores 3 extra intermediate variables the size of the image compared to FISTA. Only gradient restart scheme is implemented for now.

References:

A.B. Taylor, J.M. Hendrickx, F. Glineur, "Exact worst-case performance of first-order algorithms for composite convex optimization," Arxiv:1512.07516, 2015, SIAM J. Opt. 2017 [http://doi.org/10.1137/16m108104x]

Kim, D., & Fessler, J. A. (2018). Adaptive Restart of the Optimized Gradient Method for Convex Optimization. Journal of Optimization Theory and Applications, 178(1), 240–263. [https://doi.org/10.1007/s10957-018-1287-4]

Required Arguments

A - forward operator

OR

AHA - normal operator (as a keyword argument)

Optional Keyword Arguments

AHA - normal operator is optional if A is supplied

reg::AbstractParameterizedRegularization - regularization term

normalizeReg::AbstractRegularizationNormalization - regularization normalization scheme; options are NoNormalization(), MeasurementBasedNormalization(), SystemMatrixBasedNormalization()

rho::Real - step size for gradient step

normalize_rho::Bool - normalize step size by the largest eigenvalue of AHA

theta::Real - parameter for predictor-corrector step

sigma_fac::Real - parameter for decreasing γ-momentum ∈ [0,1]

relTol::Real - tolerance for stopping criterion

iterations::Int - maximum number of iterations

restart::Symbol - :none, :gradient options for restarting

verbose::Bool - print residual in each iteration

See also createLinearSolver, solve.

SplitBregman

RegularizedLeastSquares.SplitBregman — Type

SplitBregman(A; AHA = A'*A, reg = L1Regularization(zero(eltype(AHA))), normalizeReg = NoNormalization(), precon = Identity(), rho = 1.e2absTol = eps(), relTol = eps(), tolInner = 1.e-6, iterations::Int = 10, iterationsInner::Int = 50, iterationsCG::Int = 10)

Creates a SplitBregman object for the forward operator A.

Required Arguments

A - forward operator

Optional Keyword Arguments

AHA - normal operator is optional if A is supplied

reg::AbstractParameterizedRegularization - regularization term

normalizeReg::AbstractRegularizationNormalization - regularization normalization scheme; options are NoNormalization(), MeasurementBasedNormalization(), SystemMatrixBasedNormalization()

precon - preconditionner for the internal CG algorithm

rho::Real - weights for condition on regularized variables; can also be a vector for multiple regularization terms

absTol::Float64 - absolute tolerance for stopping criterion

relTol::Float64 - relative tolerance for stopping criterion

tolInner::Float64 - tolerance for CG stopping criterion

iterations::Int - maximum number of iterations

iterationsInner::Int - maximum number of inner iterations

iterationsCG::Int - maximum number of CG iterations

See also createLinearSolver, solve.

Miscellaneous Functions

RegularizedLeastSquares.linearSolverList — Function

Return a list of all available linear solvers

RegularizedLeastSquares.createLinearSolver — Function

createLinearSolver(solver::AbstractLinearSolver, A; kargs...)

This method creates a solver. The supported solvers are methods typically used for solving regularized linear systems. All solvers return an approximate solution to Ax = b.

TODO: give a hint what solvers are available

RegularizedLeastSquares.SolverInfo — Type

The Solver Info type is used to obtain aditional information of linear solvers and their iteration process.

Fields

convMeas::Vector{Float64} - Vector with convergence parameters

nrmse::Vector{Float64} - NRMSD of the current iterate w.r.t some reference (if provided)

x_ref::Vector{T} - Reference for computing NRMSD of iterates

x_iter::Vector{Vector{T}} - Vector for storing iterates

store_solutions::Bool - specification whether iterates are to be stored

the content of convMeas is specific to each solver. It can be found in the documentation of the corresping solve-method.

RegularizedLeastSquares.applicableSolverList — Function

applicable(args...)

list all solvers that are applicable to the given arguments. Arguments are the same as for isapplicable without the solver type.

See also isapplicable, linearSolverList.

RegularizedLeastSquares.isapplicable — Function

isapplicable(solverType::Type{<:AbstractLinearSolver}, A, x, reg)

return true if a solver of type solverType is applicable to system matrix A, data x and regularization terms reg.